requirements.txt
flask
pytesseract
pdf2image
Pillow
opencv-python
PyPDF2
flask-wtf
WTForms
email-validator
werkzeug
google-auth
SQLAlchemy
psycopg2-binary
